GEP-44 Gameboy - CPU Instrs 테스트 09 통과
 언더플로우로 인한 Z 플래그가 int 기준이었음 .

+++ gameboy/src/gameboy/instruction_set/arithmetic_instruction.cpp
@@ -25,7 +25,7 @@ void GameboyCPU::commonSubSetFlag( BYTE origin_value, BYTE sub_value, BYTE carry
 	setFlagC(  origin_value < sub_value_sum ); // 내림 당함!
 	setFlagH( ( origin_value & 0x0fu ) < ( ( sub_value & 0x0fu ) + ( carry ) ) );
 	setFlagN( true );
-	setFlagZ( ( origin_value - sub_value - carry ) == 0 );
+	setFlagZ( static_cast<BYTE>( origin_value - sub_value - carry ) == 0 );
 }
 
 void GameboyCPU::commonBitSetFlag()
